+++ obj/src/index.d.ts
@@ -2,4 +2,5 @@
 export * from './connect';
 export * from './clients';
 export * from './services';
+export * from './auth';
 export * from './build';