+++ src/clients/RestClient.ts
@@ -201,7 +201,7 @@ export abstract class RestClient implements IOpenable, IConfigurable, IReference
 
             try {
                 this._uri = connection.getUri();
-                let restify = require('restify');
+                let restify = require('restify-clients');
                 this._client = restify.createJsonClient({ 
                     url: this._uri, 
                     connectTimeout: this._connectTimeout,