+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Remote Procedure Calls for Pip.Services in Node.js Changelog
 
+## <a name="3.1.0"></a> 3.1.0 (2019-02-04)
+
+### Features
+* **auth** Authentication managers: BasicAuthManager, OwnerAuthManager and RoleAuthManager
+* **connect** HttpConnectionResolver now supports HTTPS certificates
+* **services** HttpEndpoint now supports HTTPS protocol and maintenance mode
+* **services** Added RestOperations, AboutOperations, HeartbeatOperations and StatusOperations
+* **services** Added registerRouteWithAuth and registerMiddleware methods in RestService
+
 ## <a name="2.9.0"></a> 2.9.0 (2018-04-05)
 
 ### Features